requirements.txt
av==16.0.1
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
streamlit==1.52.1
streamlit-webrtc==0.64.5
torch==2.9.1+cu126
torchvision==0.24.1+cu126
ultralytics==8.3.235